고

+++ src/main/java/com/hts/market/domain/board/entity/BrdEntity.java
@@ -13,5 +13,5 @@ public class BrdEntity {
     private Integer brdViews;
     private Long brdMemNo;
     private LocalDateTime brdCreatedAt;
-    private Integer brdCateNo;
+    private Long brdCateNo;
 }
